[API] (Add full CRUD endpoints)
- Added full CRUD endpoints for Property, PropertyImage, and PropertyTrace
- Integrated validation using FluentValidation
- Connected related entities (Address, Owner) where needed

+++ Million/Million.Application/Database/Commands/Owner/DeleteOwnerCommand/DeleteOwnerCommand.cs
@@ -1,4 +1,6 @@
-﻿using Million.Domain.Entities.Owner;
+﻿using Microsoft.EntityFrameworkCore;
+using Million.Application.Database.Commands.Address.DeleteAddressCommand;
+using Million.Domain.Entities.Owner;
 
 namespace Million.Application.Database.Commands.Owner.DeleteOwnerCommand
 {
@@ -6,12 +8,14 @@ namespace Million.Application.Database.Commands.Owner.DeleteOwnerCommand
     /// Command to delete an owner from the database.
     /// </summary>
     /// <param name="db"></param>
-    public class DeleteOwnerCommand(IDataBaseService db) : IDeleteOwnerCommand
+    public class DeleteOwnerCommand(IDataBaseService db, IDeleteAddressCommand deleteAddress) : IDeleteOwnerCommand
     {
         /// <summary>
-        /// 
+        /// Constructor.
         /// </summary>
         private readonly IDataBaseService _db = db;
+        private readonly IDeleteAddressCommand _deleteAddress = deleteAddress;
+
         /// <summary>
         /// This method is used to delete the owner info in the database.
         /// </summary>
@@ -19,8 +23,14 @@ namespace Million.Application.Database.Commands.Owner.DeleteOwnerCommand
         /// <returns></returns>
         public async Task<bool> Execute(DeleteOwnerModel model)
         {
-            OwnerEntity? ownerEntity = await _db.Owners.FindAsync(model.IdOwner);
+            OwnerEntity? ownerEntity = await _db.Owners.FirstOrDefaultAsync(p => p.IdOwner == model.IdOwner);
             if (ownerEntity == null) return false;
+
+            Task<bool> resultAddress = _deleteAddress.Execute(new DeleteAddressModel
+            {
+                IdAddress = ownerEntity.IdAddress
+            });
+            if (!resultAddress.Result) return false;
 
             _db.Owners.Remove(ownerEntity);
             await _db.SaveAsync();
